After Init Debug

+++ src/main/java/IDataLoader.java
@@ -4,7 +4,7 @@ public interface IDataLoader {
     boolean loadData(String filepath);
 
 
-    public List<List<Float>> getTrain_data();
-    public List<Float> getTrain_labels();
+    public List<List<Double>> getTrain_data();
+    public List<Double> getTrain_labels();
 
 }
